+++ deepi.php
@@ -6,9 +6,9 @@ Description: Upgrade your site's "lexical search" to Deepi's "Conceptual Search"
 Plugin Header Comment 
 Text Domain: deepi
 Domain Path: /lang
-Version: 1.4.7 beta
+Version: 1.4.9 beta
 Requires at least: 5.6
-License: MIT
+License: gpl-3.0
 Author: Mohsen Nasr 
 */
 if(!defined('ABSPATH')){